ajsute e validaçoes

+++ app/Http/Controllers/FornecedorController.php
@@ -4,11 +4,34 @@ namespace App\Http\Controllers;
 
 use App\Models\Event;
 use App\Models\Fornecedor;
+use App\Models\Servico;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class FornecedorController extends Controller
 {
+    public function validar(Request $request)
+    {
+            // Regras de validação
+        $rules = [
+            'nome' => 'required',
+            'cnpj' => 'required|size:14|unique:fornecedores,cnpj', // Supondo que o cnpj tenha 14 dígitos e seja único na tabela fornecedores.
+            'celular' => 'required',
+        ];
+
+        // Mensagens personalizadas
+        $messages = [
+            'nome.required' => 'O campo nome é obrigatório.',
+            'cnpj.required' => 'O campo CNPJ é obrigatório.',
+            'cnpj.size' => 'O CNPJ deve ter exatamente 14 dígitos.',
+            'cnpj.unique' => 'O CNPJ já está registrado.',
+            'celular.required' => 'O campo celular é obrigatório.',
+        ];
+
+        // Validar os campos
+        $request->validate($rules, $messages);
+    }
+
     public function index()
     {
         $fornecedores = Fornecedor::all();
@@ -24,11 +47,12 @@ class FornecedorController extends Controller
 
     public function store(Request $request)
     {
+        $this->validar($request);
         $fornecedor = new Fornecedor;
-        $fornecedor->nome = $request->nome;
-        $fornecedor->cnpj = $request->cnpj;
+        $fornecedor->nome    = $request->nome;
+        $fornecedor->cnpj    = $request->cnpj;
         $fornecedor->celular = $request->celular;
-        $fornecedor->iduser = Auth::user()->id; // supondo que o usuário esteja autenticado
+        $fornecedor->iduser  = Auth::user()->id; // supondo que o usuário esteja autenticado
         $fornecedor->save();
 
         return redirect('/fornecedor');
@@ -44,6 +68,7 @@ class FornecedorController extends Controller
 
     public function update(Request $request, $idfornecedor)
     {
+        $this->validar($request);
         $fornecedor = Fornecedor::findOrFail($idfornecedor);
         $fornecedor->nome = $request->nome;
         $fornecedor->cnpj = $request->cnpj;
@@ -56,6 +81,12 @@ class FornecedorController extends Controller
 
     public function destroy($idfornecedor)
     {
+      
+        $hasServicos = Servico::where('idfornecedor', $idfornecedor)->exists();
+        if($hasServicos){
+            return redirect('/fornecedor')->with('error', 'Não é possível excluir o fornecedor, pois existem serviços vinculados a ele.');
+        }  
+
         $fornecedor = Fornecedor::findOrFail($idfornecedor);
         $fornecedor->delete();
