+++ app/Http/Controllers/ServicoController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Event;
 use App\Models\Fornecedor;
 use App\Models\Servico;
 use Illuminate\Http\Request;
@@ -9,6 +10,28 @@ use Illuminate\Support\Facades\Auth;
 
 class ServicoController extends Controller
 {
+    public function validar(Request $request)
+    {
+         // Regras de validação
+        $rules = [
+            'nome' => 'required',
+            'descricao' => 'required|max:255', // Assumindo que você quer limitar a descrição em 255 caracteres. Ajuste se necessário.
+            'idfornecedor' => 'required|exists:fornecedores,id', // Verifica se o fornecedor informado existe na tabela 'fornecedores'.
+        ];
+
+        // Mensagens personalizadas
+        $messages = [
+            'nome.required' => 'O campo nome é obrigatório.',
+            'descricao.required' => 'O campo descrição é obrigatório.',
+            'descricao.max' => 'A descrição deve ter no máximo 255 caracteres.',
+            'idfornecedor.required' => 'O campo fornecedor é obrigatório.',
+            'idfornecedor.exists' => 'O fornecedor selecionado não é válido.',
+        ];
+
+        // Validar os campos
+        $request->validate($rules, $messages);
+    }
+    
     public function index()
     {
         $servicos = Servico::all();
@@ -25,6 +48,7 @@ class ServicoController extends Controller
 
     public function store(Request $request)
     {
+        $this->validar($request);
         $servico = new Servico;
         $servico->nome = $request->nome;
         $servico->descricao = $request->descricao;
@@ -46,6 +70,7 @@ class ServicoController extends Controller
 
     public function update(Request $request, $idservico)
     {
+        $this->validar($request);
         $servico = Servico::findOrFail($idservico);
         $servico->nome = $request->nome;
         $servico->descricao = $request->descricao;
@@ -58,6 +83,12 @@ class ServicoController extends Controller
 
     public function destroy($idservico)
     {
+        $existeEvento = Event::where('idservico', '=', $idservico)->exists();
+        if($existeEvento){
+            return redirect('/locais')->with('msg', 'Não é possível excluir o serviço, pois existem eventos vinculados a ele.');
+        }
+
+
         $servico = Servico::findOrFail($idservico);
         $servico->delete();
